requirements.txt
python-telegram-bot==20.3
requests==2.32.3
google-generativeai==0.3.2
python-dotenv==1.0.0
databases==0.8.0
sqlalchemy==2.0.27
mindee==3.12.0
aiosqlite==0.19.0
aiohttp==3.9.1
    